Implement get_close_matches
If the user makes an error in entering the word then instead of giving an error right away the program tries to find the closest word in the dictionary if possible.
If it cannot find a word with a match ratio of .8 of higher then it will give an error.